extract out variant  lookup
